requirements.txt
streamlit
opencv-python-headless
numpy
pyzbar
Pillow
pandas
pytesseract